toppartitions: Pass the actual list_size to the gather method instead of the capacity
Currently we're passing capacity as list_size to the gather method instead of
the actual list_size. It's done this way because every shard contains up to
list_size entries in its topk object, resulting in a maximum of list_size * smp::count
entries in total. All these entries will be copied to a separate topk object which is created
with capacity equals to the list_size passed to the gather method.
Passing capacity, which is naturally bigger than list_size, increases the precision of the calculated cardinality,
since it makes possible to copy more entries to the results object which will be returned to the storage_service,
where the results will later be truncated into a list of list_size entries.
However, this is problematic, since we have no guarantee that capacity >= (list_size * smp::count)
and also because passing capacity as list_size might be confusing and because it copies more data than necessary
into the results object.
This patch fixes this by passing the actual list_size to the gather method while making sure that the
topk object which is receiving the entries has a capacity of (list_size * smp::count).
This patch also increases the precision of the final calculated cardinality, by summing the cardinalities
calculated in each shard instead of relying on the cardinality of the results object, which might be wrong
depending on the capacity used.