toppartitions: Add toppartitions_per_shard API
This patch adds a new toppartitions_per_shard API which works similarly
to the already existent toppartitions API, but instead of returning the
top partitions during the last duration ms it will return the top partitions
ever since scylla was bootstrapped.
The only differences in usage to the already existent API is that it isn't
blocking, it doesn't take duration as a parameter and it returns list_size
partitions per shard instead of only list_size in total.